# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TranscribeAudioBody {
  /** Base64-encoded audio data */
  audio: string;
  /** Audio format (e.g. wav, m4a, webm) */
  format?: string;
}

export interface TranscribeAudioResponse {
  transcript: string;
}

export interface GenerateReplyBody {
  /** The transcribed voice input from the user */
  transcript: string;
  /** Optional context (the original message being replied to) */
  context?: string;
  /** Optional tone hint (friendly, formal, casual) */
  tone?: string;
}

export interface GenerateReplyResponse {
  reply: string;
}

export interface VoiceReplyError {
  error: string;
}