# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Transcribe audio to text
 */
export const TranscribeAudioBody = zod.object({
  audio: zod.string().describe("Base64-encoded audio data"),
  format: zod
    .string()
    .optional()
    .describe("Audio format (e.g. wav, m4a, webm)"),
});

export const TranscribeAudioResponse = zod.object({
  transcript: zod.string(),
});

/**
 * @summary Generate a polished reply from voice intent
 */
export const GenerateReplyBody = zod.object({
  transcript: zod
    .string()
    .describe("The transcribed voice input from the user"),
  context: zod
    .string()
    .optional()
    .describe("Optional context (the original message being replied to)"),
  tone: zod
    .string()
    .optional()
    .describe("Optional tone hint (friendly, formal, casual)"),
});

export const GenerateReplyResponse = zod.object({
  reply: zod.string(),
});