# lib/api-zod/src/generated/types/generateReplyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateReplyBody {
  /** The transcribed voice input from the user */
  transcript: string;
  /** Optional context (the original message being replied to) */
  context?: string;
  /** Optional tone hint (friendly, formal, casual) */
  tone?: string;
}