# lib/api-zod/src/generated/types/transcribeAudioBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeAudioBody {
  /** Base64-encoded audio data */
  audio: string;
  /** Audio format (e.g. wav, m4a, webm) */
  format?: string;
}